PHP Code Sample
Php_sample is part of the project vTigerCRM I participated in when I
worked at Sibo. vTigerCRM is an open source business system, and my job
was to add some security control to the system. Import the file 1.sql
to the mysql database, and overlap the files Authenticate.php and
Login.php.